+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Change log for SharePointDsc
+## Unreleased
+
+* New resource: SPSearchCrawlerImpactRule
 
 ## 1.6