+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Change log for SharePointDsc
 
-## Unreleased
+## 1.6
 
-* Added new resource SPWebApplicationExtension
+* Updated SPWebApplication to allow Claims Authentication configuration
 * Updated documentation in regards to guidance on installing binaries from
   network locations instead of locally
 * New resources: SPFarmPropertyBag
@@ -23,6 +23,9 @@
 * Improved runtime of SPSearchTopology by streamlining wait processes
 * Fixed bug with SPSearchServiceApp that would throw an error about SDDL string
 * Improved output of test results for AppVeyor and VS Code based test runs
+* Fixed issue with SPWebAppPolicy if OS language is not En-Us
+* Added SPFarm resource, set SPCreateFarm and SPJoinFarm as deprecated to be
+  removed in version 2.0
 
 ## 1.5